Central component -JSON Formatter
